requirements.txt
# Web & API
fastapi
uvicorn[standard]
python-multipart
pyyaml

# OCR & PDF
paddlepaddle==3.2.2
paddleocr
pdf2image
Pillow

# NLP & Embeddings
sentence-transformers
transformers==4.41.2
accelerate==0.31.0

# Vector DB & Search
chromadb
faiss-cpu

# Progress bar
tqdm

torch
